+++ Terminal.Gui/View/View.cs
@@ -71,6 +71,16 @@ public partial class View : IDisposable, ISupportInitializeNotification
             DisposeAdornments ();
             DisposeScrollBars ();
 
+            if (Application.MouseGrabView == this)
+            {
+                Application.UngrabMouse ();
+            }
+
+            if (Application.WantContinuousButtonPressedView == this)
+            {
+                Application.WantContinuousButtonPressedView = null;
+            }
+
             for (int i = InternalSubViews.Count - 1; i >= 0; i--)
             {
                 View subview = InternalSubViews [i];
