+++ Tests/UnitTests/Application/Mouse/ApplicationMouseEnterLeaveTests.cs
@@ -53,7 +53,7 @@ public class ApplicationMouseEnterLeaveTests
             ScreenPosition = mousePosition
         };
 
-        Application._cachedViewsUnderMouse.Clear ();
+        Application.CachedViewsUnderMouse.Clear ();
 
         try
         {
@@ -82,8 +82,8 @@ public class ApplicationMouseEnterLeaveTests
         List<View> currentViewsUnderMouse = new ();
         var mouseEvent = new MouseEventArgs ();
 
-        Application._cachedViewsUnderMouse.Clear ();
-        Application._cachedViewsUnderMouse.Add (view);
+        Application.CachedViewsUnderMouse.Clear ();
+        Application.CachedViewsUnderMouse.Add (view);
 
         try
         {
@@ -117,7 +117,7 @@ public class ApplicationMouseEnterLeaveTests
         Application.Top.Add (view1);
         Application.Top.Add (view2);
 
-        Application._cachedViewsUnderMouse.Clear ();
+        Application.CachedViewsUnderMouse.Clear ();
 
         try
         {
@@ -205,7 +205,7 @@ public class ApplicationMouseEnterLeaveTests
         List<View> currentViewsUnderMouse = new ();
         var mouseEvent = new MouseEventArgs ();
 
-        Application._cachedViewsUnderMouse.Clear ();
+        Application.CachedViewsUnderMouse.Clear ();
 
         try
         {
@@ -244,7 +244,7 @@ public class ApplicationMouseEnterLeaveTests
         Application.Top.Add (view1);
         Application.Top.Add (view2);
 
-        Application._cachedViewsUnderMouse.Clear ();
+        Application.CachedViewsUnderMouse.Clear ();
 
         try
         {
@@ -360,7 +360,7 @@ public class ApplicationMouseEnterLeaveTests
         view1.Add (subView);
         Application.Top.Add (view1);
 
-        Application._cachedViewsUnderMouse.Clear ();
+        Application.CachedViewsUnderMouse.Clear ();
 
         try
         {
